Add an OtherCommitters class
This class represents the committer or committers for a series of
commits. It will be used to determine whether a pull request has had any
human interaction, or whether it is just a bot's pull request. If it's
just a bot then it can close the pull request if there's a newer one,
otherwise it needs to mention the humans that are involved to see if the
pull request is still relevant.