Reintroduce PullRequest struct
Even though this is now just representing a number it's still useful to
have it around as it models a distinct part of this problem. In the
future we will want to expand this so we can access other parts of the
pull request such as the url, so having the class will be useful for
that.